+++ test/Audit.IntegrationTest/IntegrationTests.cs
@@ -62,94 +62,7 @@ namespace Audit.IntegrationTest
                 Assert.AreEqual("px", x.FilenamePrefix);
             }
 
-#if NET461 || NETCOREAPP2_0 || NETCOREAPP3_0 || NET5_0_OR_GREATER
-            [Test]
-            [Category("PostgreSQL")]
-            public void Test_PostgreDataProvider_FluentApi()
-            {
-                var x = new PostgreSql.Providers.PostgreSqlDataProvider(_ => _
-                    .ConnectionString("c")
-                    .DataColumn("dc")
-                    .IdColumnName("id")
-                    .LastUpdatedColumnName("lud")
-                    .Schema("sc")
-                    .TableName("t")
-                    .CustomColumn("c1", ev => 1)
-                    .CustomColumn("c2", ev => 2));
-                Assert.AreEqual("c", x.ConnectionStringBuilder(null));
-                Assert.AreEqual("dc", x.DataColumnNameBuilder(null));
-                Assert.AreEqual("id", x.IdColumnNameBuilder(null));
-                Assert.AreEqual("lud", x.LastUpdatedDateColumnNameBuilder(null));
-                Assert.AreEqual("sc", x.SchemaBuilder(null));
-                Assert.AreEqual("t", x.TableNameBuilder(null));
-                Assert.AreEqual(2, x.CustomColumns.Count);
-                Assert.AreEqual("c1", x.CustomColumns[0].Name);
-                Assert.AreEqual(1, x.CustomColumns[0].Value.Invoke(null));
-                Assert.AreEqual("c2", x.CustomColumns[1].Name);
-                Assert.AreEqual(2, x.CustomColumns[1].Value.Invoke(null));
-            }
 
-            [Test]
-            [Category("PostgreSQL")]
-            public void Test_PostgreDataProvider_FluentApiBuilder()
-            {
-                var x = new PostgreSql.Providers.PostgreSqlDataProvider(_ => _
-                    .ConnectionString(ev => "c")
-                    .DataColumn(ev => "dc")
-                    .IdColumnName(ev => "id")
-                    .LastUpdatedColumnName(ev => "lud")
-                    .Schema(ev => "sc")
-                    .TableName(ev => "t")
-                    .CustomColumn("c1", ev => 1)
-                    .CustomColumn("c2", ev => 2));
-                Assert.AreEqual("c", x.ConnectionStringBuilder(null));
-                Assert.AreEqual("dc", x.DataColumnNameBuilder(null));
-                Assert.AreEqual("id", x.IdColumnNameBuilder(null));
-                Assert.AreEqual("lud", x.LastUpdatedDateColumnNameBuilder(null));
-                Assert.AreEqual("sc", x.SchemaBuilder(null));
-                Assert.AreEqual("t", x.TableNameBuilder(null));
-                Assert.AreEqual(2, x.CustomColumns.Count);
-                Assert.AreEqual("c1", x.CustomColumns[0].Name);
-                Assert.AreEqual(1, x.CustomColumns[0].Value.Invoke(null));
-                Assert.AreEqual("c2", x.CustomColumns[1].Name);
-                Assert.AreEqual(2, x.CustomColumns[1].Value.Invoke(null));
-            }
-
-            [Test]
-            [Category("PostgreSQL")]
-            public void PostgreSql()
-            {
-                SetPostgreSqlSettings();
-                TestUpdate();
-                TestInsert();
-                TestDelete();
-            }
-
-            [Test]
-            [Category("PostgreSQL")]
-            public async Task PostgreSqlAsync()
-            {
-                SetPostgreSqlSettings();
-                await TestUpdateAsync();
-            }
-
-            public void SetPostgreSqlSettings()
-            {
-                Audit.Core.Configuration.Setup()
-                    .UsePostgreSql(config => config
-                        .ConnectionString(_ => AzureSettings.PostgreSqlConnectionString)
-                        .TableName("event")
-                        .IdColumnName(_ => "id")
-                        .DataColumn("data", Audit.PostgreSql.Configuration.DataType.JSONB)
-                        .LastUpdatedColumnName("updated_date")
-                        .CustomColumn("event_type", ev => ev.EventType)
-                        .CustomColumn("some_date", ev => DateTime.UtcNow)
-                        .CustomColumn("some_null", ev => null))
-                    .WithCreationPolicy(EventCreationPolicy.InsertOnStartReplaceOnEnd)
-                    .ResetActions();
-            }
-
-#endif
 
             [Test]
             [Category("EF")]
