+++ test/Run-Tests.ps1
@@ -48,6 +48,7 @@ StartDotnetUnitTests 'Audit.Elasticsearch.UnitTest' 'Elasticsearch';
 StartDotnetUnitTests 'Audit.Kafka.UnitTest' 'Kafka';
 StartDotnetUnitTests 'Audit.MongoDB.UnitTest' 'MongoDB';
 StartDotnetUnitTests 'Audit.MySql.UnitTest' 'MySql';
+StartDotnetUnitTests 'Audit.PostgreSql.UnitTest' 'PostgreSql';
 # StartDotnetUnitTests 'Audit.IntegrationTest' 'AzureStorage' '--filter=TestCategory=AzureBlob|TestCategory=AzureStorageBlobs|TestCategory=AzureTables';
 # StartDotnetUnitTests 'Audit.AzureStorageTables.UnitTest' 'AzureTables';
 StartDotnetUnitTests 'Audit.IntegrationTest' 'Elasticsearch' '--filter=TestCategory=Elasticsearch';
